+++ FranquiciaApi/FranquiciaApi/src/main/java/com/api/FranquiciaApi/Franquicia/infraestructure/drivenadapter/mysql/model/ProductoEntity.java
@@ -15,9 +15,9 @@ import lombok.Setter;
 public class ProductoEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
     private String nombre;
-    private int stock;
+    private Integer stock;
 
     @ManyToOne
     @JoinColumn(name = "sucursal_id")
